Atualiza

+++ Medicina.Application/Exame/Service/EmpresaService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Medicina.Domain.Account;
 using Medicina.Domain.Account.Repository;
 using Medicina.Domain.Cadastro.Repository;
 using Microsoft.EntityFrameworkCore.Query;
@@ -9,50 +10,64 @@ namespace Medicina.Application.Exame.Service
     public class EmpresaService: IEmpresaService
     {
         private readonly IEmpresaRepository empresaRepository;
-        private readonly IUsuarioRepository usuarioRepository;
-        private readonly IMapper mapper;
+                private readonly IMapper mapper;
 
-        public EmpresaService(IEmpresaRepository empresaRepository, IUsuarioRepository usuarioRepository, IMapper mapper)
+        public EmpresaService(IEmpresaRepository empresaRepository, IMapper mapper)
         {
-            this.usuarioRepository = usuarioRepository;
+            
             this.empresaRepository = empresaRepository;
             this.mapper = mapper;
         }
 
         public async Task<EmpresaOutputDto> Criar(EmpresaInputDto dto, Guid usuarioId)
         {
+            if (await empresaRepository.AnyAsync(x => x.Cnpj == dto.Cnpj))
+                throw new Exception("Já existe uma empresa cadastrado com o mesmo CNPJ");
+
             var empresa = this.mapper.Map<Medicina.Domain.Cadastro.Empresa>(dto);
             empresa.UsuarioId = usuarioId;
+
             await this.empresaRepository.Save(empresa);
-
             return this.mapper.Map<EmpresaOutputDto>(empresa);
 
         }
 
-        public async Task<EmpresaOutputDto> Deletar(EmpresaInputDto dto)
+        public async Task<EmpresaOutputDto> Deletar(EmpresaInputDto dto, Guid usuarioId)
         {
-            var empresa = this.mapper.Map<Medicina.Domain.Cadastro.Empresa>(dto);
+            if (await empresaRepository.AnyAsync(x => x.Cnpj == dto.Cnpj))
+            {
+                var empresa = this.mapper.Map<Medicina.Domain.Cadastro.Empresa>(dto);
+                empresa.UsuarioId = usuarioId;
 
-            await this.empresaRepository.Delete(empresa);
+                await this.empresaRepository.Delete(empresa);
+                return this.mapper.Map<EmpresaOutputDto>(empresa);
+            }
+            else
+            {
+                throw new Exception("Este CNPJ não existe.");
 
-            return this.mapper.Map<EmpresaOutputDto>(empresa);
-
+            }
         }
-
-        public async Task<EmpresaOutputDto> Atualizar(EmpresaInputDto dto)
+        public async Task<EmpresaOutputDto> Atualizar(EmpresaInputDto dto, Guid usuarioId)
         {
-        
-            var empresa = this.mapper.Map<Medicina.Domain.Cadastro.Empresa>(dto);
 
-            await this.empresaRepository.Update(empresa);
+            if (await empresaRepository.AnyAsync(x => x.Cnpj == dto.Cnpj))
+            {                   
+                var empresa = this.mapper.Map<Medicina.Domain.Cadastro.Empresa>(dto);
+                empresa.UsuarioId = usuarioId;
 
-            return this.mapper.Map<EmpresaOutputDto>(empresa);
+                await this.empresaRepository.Update(empresa);
+                return this.mapper.Map<EmpresaOutputDto>(empresa);
+            }
+            else
+            {
+                throw new Exception("Este CNPJ não existe.");
+            }                      
 
         }
 
         public async Task<List<EmpresaOutputDto>> ObterTodos()
-        {
-           
+        {           
             var empresa = await this.empresaRepository.GetAll();
             return this.mapper.Map<List<EmpresaOutputDto>>(empresa);
         }
@@ -60,7 +75,6 @@ namespace Medicina.Application.Exame.Service
         public async Task<EmpresaOutputDto> ObterPorId(Guid id)
         {
             var empresa = await this.empresaRepository.Get(id);
-
             return this.mapper.Map<EmpresaOutputDto>(empresa);
 
         }
