+++ Medicina.Application/Exame/Service/FuncionarioService.cs
@@ -29,14 +29,13 @@ namespace Medicina.Application.Exame.Service
 
         public async Task<FuncionarioOutputDto> Criar(FuncionarioInputDto dto)
         {
+            if (await funcionarioRepository.AnyAsync(x => x.Cpf == dto.Cpf))
+                throw new Exception("Já existe um funcionario cadastrado com o mesmo CPF");
 
             var empresa = await this.empresaRepository.ObterTodasEmpresasPorCnpj(dto.Cnpj);
-            var empresaId = empresa.FirstOrDefault()?.Id;
-
+            var empresaId = empresa.FirstOrDefault()?.Id;   
             var funcionario = this.mapper.Map<Medicina.Domain.Cadastro.Funcionario>(dto);
-
             funcionario.EmpresaId = empresaId.Value;
-
             await this.funcionarioRepository.Save(funcionario);
 
             return this.mapper.Map<FuncionarioOutputDto>(funcionario);
@@ -46,36 +45,44 @@ namespace Medicina.Application.Exame.Service
 
         public async Task<FuncionarioOutputDto> Deletar(FuncionarioInputDto dto)
         {
-            var funcionario = this.mapper.Map<Medicina.Domain.Cadastro.Funcionario>(dto);
-
-            await this.funcionarioRepository.Delete(funcionario);
-
-            return this.mapper.Map<FuncionarioOutputDto>(funcionario);
+            if (await funcionarioRepository.AnyAsync(x => x.Cpf == dto.Cpf))
+            {
+                var funcionario = this.mapper.Map<Medicina.Domain.Cadastro.Funcionario>(dto);
+                await this.funcionarioRepository.Delete(funcionario);
+                return this.mapper.Map<FuncionarioOutputDto>(funcionario);
+            }
+            else
+            {
+                throw new Exception("Este CPF não existe.");
+            }
+           
 
         }
 
         public async Task<FuncionarioOutputDto> Atualizar(FuncionarioInputDto dto)
         {
-            var funcionario = this.mapper.Map<Medicina.Domain.Cadastro.Funcionario>(dto);
-
-            await this.funcionarioRepository.Update(funcionario);
-
-            return this.mapper.Map<FuncionarioOutputDto>(funcionario);
+            if (await funcionarioRepository.AnyAsync(x => x.Cpf == dto.Cpf))
+            {
+                var funcionario = this.mapper.Map<Medicina.Domain.Cadastro.Funcionario>(dto);
+                await this.funcionarioRepository.Update(funcionario);
+                return this.mapper.Map<FuncionarioOutputDto>(funcionario);
+            }
+            else
+            {
+                throw new Exception("Este CPF não existe.");
+            }                       
 
         }
 
         public async Task<List<FuncionarioOutputDto>> ObterTodos()
         {
-
             var funcionario = await this.funcionarioRepository.GetAll();
-
             return this.mapper.Map<List<FuncionarioOutputDto>>(funcionario);
         }
 
         public async Task<FuncionarioOutputDto> ObterPorId(Guid id)
         {
             var funcionario = await this.funcionarioRepository.Get(id);
-
             return this.mapper.Map<FuncionarioOutputDto>(funcionario);
 
         }
